add ut tests
